backend/requirements.txt
fastapi==0.115.12
langchain_core==0.3.51
langchain_google_community==2.0.7
langchain_google_genai==2.1.2
langchain_text_splitters==0.3.8
pinecone==6.0.2
pydantic==2.11.3
python-dotenv==1.1.0
uvicorn==0.34.0
